perpdex/paradex/requirements.txt
# Paradex Trading Bot Dependencies

# Official Paradex Python SDK
paradex-py>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests (if needed for custom API calls)
requests>=2.31.0

# WebSocket for real-time data
websocket-client>=1.6.0

# Ethereum wallet integration (required by paradex-py)
web3>=6.0.0
eth-account>=0.9.0
